Refactor context and index services out of main
