+++ app/context/__init__.py
@@ -1,0 +1,21 @@
+from .service import (
+    context_retrieve_service,
+    context_snapshot_create_service,
+    context_snapshot_get_service,
+    index_rebuild_incremental_service,
+    index_rebuild_service,
+    index_status_service,
+    recent_list_service,
+    search_service,
+)
+
+__all__ = [
+    "context_retrieve_service",
+    "context_snapshot_create_service",
+    "context_snapshot_get_service",
+    "index_rebuild_incremental_service",
+    "index_rebuild_service",
+    "index_status_service",
+    "recent_list_service",
+    "search_service",
+]
